backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "queueflow-backend"
version = "0.1.0"
description = "Multi-site Queue Orchestration Platform"
license = { text = "Apache-2.0" }
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    # Database (async)
    "sqlalchemy>=2.0.30",
    "alembic>=1.13.0",
    "asyncpg>=0.29.0",
    # Database (sync – for Celery workers and migrations)
    "psycopg2-binary>=2.9.9",
    # Validation
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    # Cache / queue broker
    "redis>=5.0.4",
    "celery[redis]>=5.4.0",
    # Auth / security
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    # Rate limiting
    "slowapi>=0.1.9",
    # HTTP client (for JWKS fetching)
    "httpx>=0.27.0",
    # Logging
    "structlog>=24.1.0",
    # Timezone
    "pytz>=2024.1",
    # Observability
    "prometheus-fastapi-instrumentator>=6.1.0",
    "opentelemetry-api>=1.27.0",
    "opentelemetry-sdk>=1.27.0",
    "opentelemetry-instrumentation-fastapi>=0.49b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.49b0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.27.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "anyio>=4.4.0",
    # Linting / formatting
    "ruff>=0.4.0",
    "black>=24.4.0",
    "mypy>=1.10.0",
    # Type stubs
    "types-pytz",
    "types-passlib",
    # Pre-commit
    "pre-commit>=3.7.0",
    # Test DB
    "aiosqlite>=0.20.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

# ── Ruff ────────────────────────────────────────────────────────────────────
[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "UP", "N", "S", "B", "A", "C4", "PT", "SIM"]
ignore = ["S101", "S105", "S106", "B008", "N818", "A002", "S311"]

[tool.ruff.per-file-ignores]
"tests/*" = ["S", "B", "N"]
"scripts/*" = ["S", "T201"]

# ── Black ────────────────────────────────────────────────────────────────────
[tool.black]
line-length = 100
target-version = ["py311"]

# ── Mypy ─────────────────────────────────────────────────────────────────────
[tool.mypy]
python_version = "3.11"
strict = false
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

# ── Pytest ────────────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::pytest.PytestUnraisableExceptionWarning",
]
